Identity v3 as default - Step 1
The goal is to makes identity v3 as the default version.
Effectively since Openstack Kilo, Keystone engine is by default version 3 while
still providing service to API v2.0.

Targetting a configuration where the endpoints for v3 don't have to be created
along the default v2.0 ones which are not deprecated yet.

This patch simplifies the identity version class building.