Fix introspection service lookup
Upstream, the service is registered as "baremetal-introspection"
as the service type name. Setting the service type name manually
produced resulted in an error in set_api_path because it could not
find a supported version.

This patch changes the type from "introspection" to
"baremetal-introspection". It also explicitly sets the api path to
v1 because the version reported by the introspection api is 1.6
which leads Fog::OpenStack.get_supported_version to throw an error.