removed extraneous tenant_id from quota_set in update_quota
In both the Compute and Volume update_quota actions, update_quota
was adding tenant_id to the quota_set hash. It doesn't belong there.
tenant_id is already in the URL and it doesn't go into the quota
hash. For the Compute operation, adding it to quota_set causes it
to fail. For Volume, it doesn't fail, but it still doesn't belong there.