Swift: support POST for objects.
When an object is saved, fog should attempt to issue a POST, as
opposed to always reuploading the object. The patch adds a @dirty
instance variable to the File model to keep track of whether the
object needs to be PUT or if a POST should be issues. To support the
model, a post_object request is also added. The expected return code
is 202.

Fixes: #289