Allow more than 100 results post-Liberty.
The Liberty release of Ceilometer introduced the idea of a limit
variable to tell the API how many results you want. By default
this is set to 100 in the server config. The onus is on the client
to select how many results it wants.

I propose we default to a very high number.